Restructure directory hierarchy [refactoring]
